Moving some components inside the verification package
