+++ vector/src/main/java/im/vector/app/features/settings/devices/v2/DevicesViewState.kt
@@ -19,6 +19,7 @@ package im.vector.app.features.settings.devices.v2
 import com.airbnb.mvrx.Async
 import com.airbnb.mvrx.MavericksState
 import com.airbnb.mvrx.Uninitialized
+import im.vector.app.features.settings.devices.v2.verification.CurrentSessionCrossSigningInfo
 
 data class DevicesViewState(
         val currentSessionCrossSigningInfo: CurrentSessionCrossSigningInfo = CurrentSessionCrossSigningInfo(),
